verify add to cart

+++ src/test/java/com/automation/pages/web/WebProductListPage.java
@@ -75,14 +75,14 @@ public class WebProductListPage extends WebBasePage {
 
     public boolean isPriceHighToLowSorted() {
 
-        List<Long> newPriceList = new ArrayList<>();
+        List<Double> newPriceList = new ArrayList<>();
 
         for (WebElement price : priceList) {
             String str = price.getText();
-            newPriceList.add(Long.valueOf(str.substring(1)));
+            newPriceList.add(Double.valueOf(str.substring(1)));
 
         }
-        List<Long> copyPriceList = new ArrayList<>(newPriceList);
+        List<Double> copyPriceList = new ArrayList<>(newPriceList);
         Collections.sort(copyPriceList);
         Collections.reverse(copyPriceList);
 
@@ -92,14 +92,14 @@ public class WebProductListPage extends WebBasePage {
 
     public boolean isPriceLowToHighSorted() {
 
-        List<Long> newPriceList = new ArrayList<>();
+        List<Double> newPriceList = new ArrayList<>();
 
         for (WebElement price : priceList) {
             String str = price.getText();
-            newPriceList.add(Long.valueOf(str.substring(1)));
+            newPriceList.add(Double.valueOf(str.substring(1)));
 
         }
-        List<Long> copyPriceList = new ArrayList<>(newPriceList);
+        List<Double> copyPriceList = new ArrayList<>(newPriceList);
         Collections.reverse(copyPriceList);
         Collections.sort(copyPriceList);
 
@@ -107,8 +107,6 @@ public class WebProductListPage extends WebBasePage {
 
     }
 
-    //    @FindBy(xpath = "//div[text()='Brand']/following-sibling::div")
-//    WebElement plusIconBrand;
     String filterPlusIcon = "//div[text()='%s']/following-sibling::div";
     @FindBy(xpath = "//input[@placeholder='Search by brands']")
     WebElement brandInput;
